+++ elmyra.ip.access.epo/CHANGES.rst
@@ -18,6 +18,7 @@ development
 - ui: improve permalink popover widget and subsystem
 - ui: basket sharing via url
 - ui: basket sharing via email: improve content details
+- ui: improve history api interaction
 
 0.24.3
 ======